+++ backend/src/main/java/meetween/backend/place/application/PlaceService.java
@@ -1,8 +1,8 @@
 package meetween.backend.place.application;
 
-import meetween.backend.place.domain.Coordinate;
+import meetween.backend.place.domain.coordinate.Coordinate;
 import meetween.backend.member.domain.MemberRepository;
-import meetween.backend.place.domain.Restaurant;
+import meetween.backend.place.domain.entity.Restaurant;
 import meetween.backend.place.domain.RestaurantRepository;
 import meetween.backend.place.dto.response.RestaurantResponse;
 import meetween.backend.place.dto.response.RestaurantSpecificResponse;
@@ -10,6 +10,7 @@ import meetween.backend.place.dto.request.PlacesByLocationRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -26,8 +27,8 @@ public class PlaceService {
         this.memberRepository = memberRepository;
     }
 
-    public RestaurantResponse getNearRestaurants(final PlacesByLocationRequest request) {
-        Coordinate coordinate = Coordinate.of(request.getLatitude(), request.getLongitude(), request.getLatitudeDelta(), request.getLongitudeDelta());
+    public RestaurantResponse getNearRestaurants(final BigDecimal latitude, final BigDecimal longitude, final BigDecimal latitudeDelta, final BigDecimal longitudeDelta) {
+        Coordinate coordinate = Coordinate.of(latitude, longitude, latitudeDelta, longitudeDelta);
 
         List<Restaurant> restaurants = restaurantRepository.findAllByLatitudeAndLongitudeBetween(
                 coordinate.getMinLatitude(),
