+++ backend/src/main/java/meetween/backend/place/domain/RestaurantRepository.java
@@ -1,5 +1,6 @@
 package meetween.backend.place.domain;
 
+import meetween.backend.place.domain.entity.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
