+++ backend/src/main/java/meetween/backend/place/infra/RestTemplateCafeRequester.java
@@ -1,6 +1,6 @@
 package meetween.backend.place.infra;
 
-import meetween.backend.place.dto.request.CafeRequest;
+import meetween.backend.place.dto.request.api.CafeRequest;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
